[feat/#181] :recycle: Refactor: 스키마 파일 생성
